kernel: compat: provide selinux_inode/selinux_cred wrappers for < 5.1
ref:
https://github.com/gregkh/linux/commit/0c6cfa622cf57109607f3416b286b6b55561a2ea
https://github.com/gregkh/linux/commit/80788c229116b28cc914d73c142f74bdee28ab57

It is a common thing on the scene to backport things, so this breaks kernel versioning
assumptions. As for those, we have to scan and check kernel source.

The following are commonly backported:
 - selinux_inode / selinux_cred
	commonly backported due to 5.10-bpf on 4.x
	https://github.com/Mi-Thorium/kernel_msm-4.19/commit/0b670b34e354c5a4f521d49c670e5fb8aeface71
	https://github.com/Mi-Thorium/kernel_msm-4.19/commit/162258a3f63e46ac1db891765af08459ae1eed85

old: https://github.com/backslashxx/KernelSU/commit/5d841fbf086c4151749a951b9b0db57fb3ae9a90

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>